+++ Alarm/AlarmListTableViewController.swift
@@ -7,8 +7,12 @@
 
 import UIKit
 
+import FirebaseDatabase
+
 class AlarmListTableViewController: UITableViewController {
 
+    var alarmArr: [Alarm] = []
+    
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -19,27 +23,34 @@ class AlarmListTableViewController: UITableViewController {
         // self.navigationItem.rightBarButtonItem = self.editButtonItem
     }
 
+    override func viewWillAppear(_ animated: Bool) {
+    
+    }
+    
     // MARK: - Table view data source
 
     override func numberOfSections(in tableView: UITableView) -> Int {
         // #warning Incomplete implementation, return the number of sections
-        return 0
+        return 1
     }
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         // #warning Incomplete implementation, return the number of rows
-        return 0
+        return self.alarmArr.count
     }
 
-    /*
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "reuseIdentifier", for: indexPath)
+        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
 
         // Configure the cell...
+        var cellConfig = cell.defaultContentConfiguration()
+        cellConfig.text = self.alarmArr[indexPath.row].alarmName
+        
+        cell.contentConfiguration = cellConfig
 
         return cell
     }
-    */
+    
 
     /*
     // Override to support conditional editing of the table view.
@@ -76,14 +87,17 @@ class AlarmListTableViewController: UITableViewController {
     }
     */
 
-    /*
+    
     // MARK: - Navigation
 
     // In a storyboard-based application, you will often want to do a little preparation before navigation
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         // Get the new view controller using segue.destination.
         // Pass the selected object to the new view controller.
+        if segue.identifier == "segueToAddAlarm" {
+            let addAlarmVC = segue.destination as? AddAlarmViewController
+            addAlarmVC?.alarmArr = self.alarmArr
+        }
     }
-    */
 
 }