+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+19 June 2024
+- password field in the login screen is now changed to a secure textfield type
+- you can now add an alarm to the database
+
 26 May 2024
 - implemented sign up